infrastructure/lambda/api/company/requirements.txt
boto3==1.34.144
PyJWT==2.8.0
cryptography==41.0.7